+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Changelog
 
+### Unreleased
+- [new] `NickClient` add support for asset deletion
+
 ### 2.7.1
 - [fix] missing name for `AssetClient.create` in nodejs